Removed vestigial UpdateContentLayout() method left over from Forms
The Forms version of ItemContentView called the UpdateContentLayout method
when laying out each RecyclerView item to ensure that all cross-platform
layout updates were applied. When the CollectionView was ported to MAUI,
that method didn't appropriately handle the changes to margins in the layout,
so a call to LayoutVirtualView was added to deal with that. Both methods
effectively do the same job, so UpdateContentLayout is now redundant.

This change removes UpdateContentLayout so we don't lay out (and in some cases
re-measure) the content unnecessarily.